Auto restart outlook and App if "folder not found"
The application would get stuck if the outlook folder had a change to it, like a new folder being added or removed. So to get around this, we need to restart the local instance of outlook to let it update, then restart the application. This is now automatic in case someone doesnt see this happen during the night